extract compaction logic into separate modules
